move propsal to common folder
